- Fixed bug in drawAnimationCommand which messed up character portraits and intro graphics - Implemented clipping rects (setClipRect and code in drawAnimationCelSprite) - needs more work - More renaming - Some more debug output (TODO: switch to debugC)
git-svn-id: https://svn.scummvm.org:4444/svn/comet@52 e395a8f9-5cf0-422e-971d-9a083d8f24e8

+++ engines/comet/animation.h
@@ -43,7 +43,7 @@ struct AnimationElement {
 
 struct AnimationCel {
 	uint16 flags;
-	byte width, height;
+	uint16 width, height;
 	uint16 dataSize;
 	byte *data;
 };
